Add flag to disable filtering of failing callbacks.
Currently if a function called by the poller fails, it's removed
from the poller configuration and not called again.
This means that transient errors can disable metrics reporting
and it cannot be re-enabled.

This change adds a `filter_misbehaving` flag (defaults to `true`)
to disable this behaviour and always call the configured measurement
functions regardless of failures.